+++ abci/client/errors.go
@@ -1,10 +1,15 @@
 package abcicli
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/cometbft/cometbft/abci/types"
 )
+
+// ErrClientStopped is returned for requests that were still in flight when
+// the client was stopped.
+var ErrClientStopped = errors.New("ABCI client is stopped")
 
 // ErrUnknownAbciTransport is returned when trying to create a client with an invalid transport option
 type ErrUnknownAbciTransport struct {